pyproject.toml
[tool.poetry]
name = "profiling_test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tg4perfetto = "^0.0.4"
protobuf = "3.19.0"
python-hunter = "^0.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
